pavex_session

Module config

Types related to SessionConfig.

Structs

InvalidTtlExtensionThreshold

Error raised when trying to create a TtlExtensionThreshold with an invalid value.

SessionCookieConfig

Configure the cookie used to store session information on the client-side.

SessionStateConfig

Configure the way session state is stored.

TtlExtensionThreshold

A ratio between 0 and 1, interpreted as a percentage of the time-to-live of a fresh session.

Enums

MissingServerState

Configure the expected behaviour when dealing with a pre-existing session (i.e. with a valid client-side cookie) that doesn’t have a corresponding record in the session store.

ServerStateCreation

Configure when the storage backend should be asked to create a new session state record.

SessionCookieKind

The kind of cookie used to store session information on the client-side.

TtlExtensionTrigger

Configure when the TTL for an existing session should be extended.