pavex_session::config

Enum TtlExtensionTrigger

#[non_exhaustive]
pub enum TtlExtensionTrigger {
    OnStateLoadsAndChanges,
    OnStateChanges,
}

Configure when the TTL for an existing session should be extended.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

OnStateLoadsAndChanges

The TTL of the current session is refreshed on every request where the server either:

Modified the session state

Loaded the server state from the storage backend

This is the default.

Performance impact

TTL refreshes are not free, as they require an additional request to the storage backend if the state was otherwise unchanged. This impacts both the total load on your storage backend (i.e. number of queries it has to handle) and the latency of the requests served by your server.

This impact can be mitigated by setting a TTL extension threshold.

OnStateChanges

The TTL of the current session is only refreshed when the session state is modified.

It doesnโt distinguish between changes to the client-side and the server-side session state when determining if the TTL should be refreshed.

Performance impact

OnStateChanges may reduce the number of requests to the storage backend compared to OnStateLoadsAndChanges, as well as improve the latency of the requests served by your server by removing a network request from the critical path. It primarily depends on the TTL extension threshold you set, if any.