pavex_session

Module errors

Errors that can occur when interacting with the session state.

Structs

ValueDeserializationError

Returned when we fail to deserialize a value stored in either the server or the client session state.

ValueSerializationError

Returned when we fail to serialize a value to be stored in either the server or the client session state.

Enums

FinalizeError

The error returned by finalize_session.

ServerGetError

The error returned by Session::get.

ServerInsertError

The error returned by Session::insert.

ServerRemoveError

The error returned by Session::remove.

SyncError

The error returned by Session::sync.

Constants

FINALIZE_ERROR_INTO_RESPONSE

A handle to add FinalizeError::into_response() as an error handler to your Pavex application.