pavex_session

Module store

Types and traits related to SessionStore.

Modules

errors

Errors that can occur when interacting with a session storage backend.

Structs

SessionRecord

A server-side session record that was retrieved from the chosen storage backend.

SessionRecordRef

A server-side session record that’s going to be stored in the chosen storage backend.

Traits

SessionStorageBackend

The interface of a session storage backend.