pavex_session

Struct Session

pub struct Session<'store> { /* private fields */ }

The current HTTP session.

Implementation notes

Not Clone

The session is a stateful object that holds the client-side and server-side state of the session, tracking all changes to both states. As a result, Session does not implement the Clone trait.

Not Send nor Sync

The session object is designed to be used within the lifetime of the request it refers to. When Pavex receives a new request, it assigns it to a specific worker thread, where all the processing for that request takes place.

Given the above, we optimized Session’s internals for single-thread usage and decided not to implement Send and Sync for it.

Implementations

impl<'store> Session<'store>

pub fn new( store: &'store SessionStore, config: &'store SessionConfig, incoming_session: Option<IncomingSession>, ) -> Self

Create a new HTTP session.

It is a continuation of the existing session if there was a valid session cookie attached to the request. It is a brand-new session otherwise.

impl Session<'_>

All the operations you can perform on the server-side state of your session.

pub async fn get<T: DeserializeOwned>( &self, key: &str, ) -> Result<Option<T>, ServerGetError>

Get the value associated with key from the server-side state.

If the value is not found, None is returned.
If the value is found, but it cannot be deserialized into the expected type, an error is returned.

If you don’t need to deserialize the value, or you’d like to handle the deserialization yourself, use get_raw instead.

pub async fn insert<T, Key>( &mut self, key: Key, value: T, ) -> Result<Option<Value>, ServerInsertError>
where T: Serialize, Key: Into<Cow<'static, str>>,

Insert a value for the given key in the server-side state.

If the state didn’t have an entry for this key, the value is inserted and None is returned.
If the state did have an entry for this key, its value is updated and the old value is returned in its raw JSON form.

The provided value is serialized to JSON prior to being stored. If the serialization fails, an error is returned. If you’d prefer to take care of the serialization yourself, use insert_raw instead.

pub async fn remove<T: DeserializeOwned>( &mut self, key: &str, ) -> Result<Option<T>, ServerRemoveError>

Remove the value associated with key from the server-side state.

If the key doesn’t exist, None is returned.

If the key exists, the removed value is returned, deserialized into the type you specify as T. If the removed value cannot be deserialized, an error is returned.

If you’re not interested in the removed value, or you don’t want to deserialize it, use remove_raw instead.

pub async fn is_empty(&self) -> Result<bool, LoadError>

Returns true if there are no values in the server-side state.

pub async fn get_raw<'a>( &'a self, key: &str, ) -> Result<Option<&'a Value>, LoadError>

Get the value associated with key from the server-side state.

pub async fn insert_raw<Key>( &mut self, key: Key, value: Value, ) -> Result<Option<Value>, LoadError>
where Key: Into<Cow<'static, str>>,

Insert a value for the given key in the server-side state.

If the state didn’t have an entry for this key, the value is inserted and None is returned.
If the state did have an entry for this key, its value is updated and the old value is returned in its raw JSON form.

The provided value must be a JSON value, which will be stored as-is, without any further manipulation. If you’d prefer to let pavex_session handle the serialization, use insert instead.

pub async fn remove_raw( &mut self, key: &str, ) -> Result<Option<Value>, LoadError>

Remove the value associated with key from the server-side state.

If the key exists, the removed value is returned.
The value is returned as it was stored in the server-side state, without any deserialization. If you want to deserialize the value as a specific type, use remove instead.

pub fn delete(&mut self)

Delete the session record from the store.

This doesn’t destroy the whole session—you must invoke Session::invalidate if that’s your goal.

pub async fn clear(&mut self) -> Result<(), LoadError>

Remove all key-value pairs from the server-side state.

This doesn’t delete the session record from the store—you must invoke Session::delete if you want to delete the record altogether.

This doesn’t invalidate the session—you must invoke Session::invalidate if you want to delete the session altogether.

pub fn cycle_id(&mut self)

Generate a new session identifier and attach it to this session. The session state is preserved on both the client-side and the server-side.

This method is useful for security reasons, as it can help prevent session fixation attacks.

pub fn invalidate(&mut self)

Invalidate the session.

The server-side session state will be marked for deletion. The client-side cookie will be removed from the client using a removal cookie.

After calling this method, the session is considered invalid and should not be used anymore. All further operations on the session will be no-ops.

pub fn is_invalidated(&self) -> bool

Check if the session has been invalidated.

See Session::invalidate for more information.

impl Session<'_>

Control when the server-side state is synchronized with the store.

pub async fn sync(&mut self) -> Result<(), SyncError>

Sync the in-memory representation of the server-side state with the store.

In most cases, you don’t need to invoke this method manually: it is done for you by finalize_session, the post-processing middleware that attaches the session cookie to the response returned to the client.

pub async fn force_load(&self) -> Result<(), LoadError>

Load the server-side state from the store. This method does nothing if the server-side state has already been loaded.

After calling this method, the server-side state will be loaded and cached in memory, so that subsequent calls to get_raw, insert_raw, and remove_raw will operate on the in-memory state.

pub async fn finalize( &mut self, ) -> Result<Option<ResponseCookie<'static>>, FinalizeError>

Sync the current server-side state with the chosen storage backend. If necessary, it returns a cookie to be attached to the outgoing response in order to sync the client-side state.

impl Session<'_>

APIs for manipulating the client-side session state.

pub fn client(&self) -> ClientSessionState<'_>

Read values from the client-side state attached to this session.

pub fn client_mut(&mut self) -> ClientSessionStateMut<'_>

Read or mutate the client-side state attached to this session.